+++ src/main/kotlin/io/github/cotrin8672/cel/content/block/vault/EnderVaultBlock.kt
@@ -2,9 +2,20 @@ package io.github.cotrin8672.cel.content.block.vault
 
 import com.simibubi.create.content.equipment.wrench.IWrenchable
 import com.simibubi.create.foundation.block.IBE
+import io.github.cotrin8672.cel.content.SharedStorageBehaviour
 import io.github.cotrin8672.cel.registry.CelBlockEntityTypes
+import io.github.cotrin8672.cel.registry.CelBlocks
+import io.github.cotrin8672.cel.registry.CelDataComponents
+import io.github.cotrin8672.cel.util.CelLang
+import io.github.cotrin8672.cel.util.StorageFrequency
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
+import net.minecraft.server.level.ServerPlayer
+import net.minecraft.world.InteractionHand
+import net.minecraft.world.ItemInteractionResult
+import net.minecraft.world.entity.LivingEntity
+import net.minecraft.world.entity.player.Player
+import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.context.BlockPlaceContext
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.Block
@@ -15,6 +26,7 @@ import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.level.block.state.StateDefinition
 import net.minecraft.world.level.block.state.properties.BlockStateProperties
 import net.minecraft.world.level.block.state.properties.EnumProperty
+import net.minecraft.world.phys.BlockHitResult
 
 class EnderVaultBlock(properties: Properties) : Block(properties), IWrenchable, IBE<EnderVaultBlockEntity> {
     companion object {
@@ -45,6 +57,49 @@ class EnderVaultBlock(properties: Properties) : Block(properties), IWrenchable, 
 
     override fun getStateForPlacement(context: BlockPlaceContext): BlockState {
         return defaultBlockState().setValue(HORIZONTAL_AXIS, context.horizontalDirection.axis)
+    }
+
+    override fun useItemOn(
+        stack: ItemStack,
+        state: BlockState,
+        level: Level,
+        pos: BlockPos,
+        player: Player,
+        hand: InteractionHand,
+        hitResult: BlockHitResult,
+    ): ItemInteractionResult {
+        val heldItem = player.getItemInHand(hand)
+        if (CelBlocks.ENDER_VAULT.isIn(heldItem)) {
+            withBlockEntityDo(level, pos) {
+                val storageFrequency = it.getBehaviour(SharedStorageBehaviour.TYPE).getFrequency()
+                if (storageFrequency.isEmpty)
+                    heldItem.remove(CelDataComponents.STORAGE_FREQUENCY)
+                else
+                    heldItem.set(CelDataComponents.STORAGE_FREQUENCY, storageFrequency)
+            }
+            if (player is ServerPlayer)
+                player.displayClientMessage(
+                    CelLang.translate("storage_frequency.set").component(),
+                    true
+                )
+            return ItemInteractionResult.SUCCESS
+        }
+
+        return ItemInteractionResult.SKIP_DEFAULT_BLOCK_INTERACTION
+    }
+
+    override fun setPlacedBy(level: Level, pos: BlockPos, state: BlockState, placer: LivingEntity?, stack: ItemStack) {
+        super.setPlacedBy(level, pos, state, placer, stack)
+        val storageFrequency = stack.getOrDefault(
+            CelDataComponents.STORAGE_FREQUENCY,
+            StorageFrequency.EMPTY
+        )
+        if (storageFrequency.isNotEmpty) {
+            withBlockEntityDo(level, pos) {
+                val behaviour = it.getBehaviour(SharedStorageBehaviour.TYPE)
+                behaviour.setStorageFrequency(storageFrequency)
+            }
+        }
     }
 
     public override fun onRemove(