+++ src/main/kotlin/io/github/cotrin8672/cel/content/item/ScopeFilterItemDecorator.kt
@@ -1,51 +1,23 @@
 package io.github.cotrin8672.cel.content.item
 
-import com.mojang.blaze3d.platform.Lighting
 import io.github.cotrin8672.cel.registry.CelDataComponents
-import net.minecraft.client.Minecraft
+import io.github.cotrin8672.cel.util.use
 import net.minecraft.client.gui.Font
 import net.minecraft.client.gui.GuiGraphics
-import net.minecraft.client.renderer.texture.OverlayTexture
-import net.minecraft.world.item.ItemDisplayContext
 import net.minecraft.world.item.ItemStack
 import net.neoforged.neoforge.client.IItemDecorator
-import thedarkcolour.kotlinforforge.neoforge.forge.use
 
 object ScopeFilterItemDecorator : IItemDecorator {
     override fun render(guiGraphics: GuiGraphics, font: Font, stack: ItemStack, xOffset: Int, yOffset: Int): Boolean {
         val storageFrequency = stack.get(CelDataComponents.STORAGE_FREQUENCY) ?: return false
         val frequencyItem = storageFrequency.stack
         if (frequencyItem.isEmpty) return false
-        guiGraphics.renderItemModel(xOffset + 8f, yOffset + 8f, 9f, frequencyItem)
-        return true
-    }
-
-    private fun GuiGraphics.renderItemModel(
-        x: Float,
-        y: Float,
-        scale: Float,
-        stack: ItemStack,
-    ) {
-        val minecraft = Minecraft.getInstance()
-        val model = minecraft.itemRenderer.getModel(stack, minecraft.level, minecraft.player, 16777216)
-        pose().use {
-            pose().translate(x, y, 160f)
-            pose().scale(scale, -scale, scale)
-            val flag = !model.usesBlockLight()
-            if (flag) Lighting.setupForFlatItems()
-
-            minecraft.itemRenderer.render(
-                stack,
-                ItemDisplayContext.GUI,
-                false,
-                pose(),
-                bufferSource(),
-                15728880,
-                OverlayTexture.NO_OVERLAY,
-                model
-            )
-            flush()
-            if (flag) Lighting.setupFor3DItems()
+        guiGraphics.pose().use {
+            translate(xOffset + 8f, yOffset + 8f, 0f)
+            scale(0.5f, 0.5f, 1f)
+            translate(-(xOffset + 8f), -(yOffset + 8f), 40f)
+            guiGraphics.renderItem(frequencyItem, xOffset, yOffset)
         }
+        return true
     }
 }